Prepare for launcher prebuilts on the host
Instead of building python3 for the host for every build using
`embedded_launcher: true`, I'm preparing to switch us to using prebuilts
from prebuilts/build-tools. Besides improving build speed, this should
let us enable the _ssl module, which requires openssl, which is not
available on platform branches.

The changes to actually support using the prebuilts will come in a later
change, as we need to get the prebuilts in the tree first.

Test: treehugger
Change-Id: Ibc99c3e41ea8c03e8d17ec38cbd25b24f7d96acf