Use a $ORIGIN/../lib rpath in extension modules
Previously, make invoked setup.py with something like:

    setup.py LDSHARED='... -Wl,-rpath,'$ORIGIN/../lib' ...'

This let the shell expand $ORIGIN to an empty string, so the rpath for
the stdlib extension modules was /../lib, which presumably meant /lib.
I don't see a way to provide a different rpath for different Python
binaries, so change the escaping so we use $ORIGIN/../lib consistently.
This directory won't exist for the extension modules in
${python3}/lib/python3.9/lib-dynload, but that's OK.

Bug: https://github.com/android/ndk/issues/1556
Change-Id: Id05e5a9513376f78e342f214422254097813dd71