Use upstream zlib's CMake build system
We need to use the "install" target to get both zlib.h and zconf.h
in the same directory, separate from the internal zlib headers.

Set the "PythonForBuild" property to "python.exe". This seems to be
necessary to build python on my local machine, but hasn't been
necessary in Kokoro's CI. I'm not sure why.

Bug: none
Test: external\python\cpython3\kokoro\kokoro_build.cmd
Change-Id: I3cd6ebb3f46b4d09cc45a6e557eb8b66d567d514